CommsIntegration/pyproject.toml
[project]
name = "commsintegration"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "dotenv>=0.9.9",
    "paho-mqtt>=2.1.0",
    "pika>=1.3.2",
    "redis>=7.1.0",
    "sentry-sdk>=2.50.0",
]
